fix : tts error fix
- 동시 출력, 요청 에러, 음성 출력 시 api 요청 금지 등으로 해결
