+++ app/src/components/game/templates/SelectRenderContent.jsx
@@ -1,52 +1,27 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React, { useEffect, useRef } from 'react';
 import TurnSelect from '../organisms/TurnSelect';
 import SubjectSelect from '../organisms/SubjectSelect';
-import useGameTurn from '../../../hooks/useGameTurn';
-import { setTempTurn } from '../../../redux/slice/TemplateTurn';
 import GameTemplate from './GameTemplate';
+import useGameSelectHandler from '../../../hooks/useGameSelectHandler';
+import useTTS from '../../../hooks/useTTS';
 
 const SelectRenderContent = ({ navigation }) => {
-  const [stage, setStage] = useState('turnSelect');
-  const [subtitleText, setSubtitleText] = useState("순서를 정해보자\n어떤 순서로 하고 싶어?");
-  const [loading, setLoading] = useState(false);
-  const [userInput, setUserInput] = useState("");
-  const dispatch = useDispatch();
-  const { sendRequest } = useGameTurn();
-  const tempTurn = useSelector((state) => state.templateTurn.tempTurn);
+  const {
+    stage,
+    subtitleText,
+    loading,
+    handleButtonPress,
+    handleSubjectPress,
+  } = useGameSelectHandler(navigation);
+  const { loading: ttsLoading, handleTTS } = useTTS();
+  const prevSubtitleText = useRef();
 
-  const handleButtonPress = async (message, turnValue, title) => {
-    setLoading(true);
-    setUserInput(title);
-    if (turnValue) {
-      dispatch(setTempTurn(turnValue));
+  useEffect(() => {
+    if (subtitleText && subtitleText !== prevSubtitleText.current) {
+      handleTTS(subtitleText);
+      prevSubtitleText.current = subtitleText;
     }
-    try {
-      const response = await sendRequest(title);
-      setSubtitleText(response.data.response[0].message);
-      setStage('subjectSelect'); // 주제 선택 화면으로 전환
-    } catch (error) {
-      console.error('Error handling turn selection:', error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleSubjectPress = async (message) => {
-    setLoading(true);
-    setUserInput(message);
-
-    // 주제 선택 버튼을 눌렀을 때만 요청을 보냅니다.
-    try {
-      const response = await sendRequest(message);
-      console.log('Response:', response);
-      setLoading(false);
-      navigation.navigate('gamescreen', { message: response.data.response[0].message });
-    } catch (error) {
-      console.error('Error handling user input:', error);
-      setLoading(false);
-    }
-  };
+  }, [subtitleText, handleTTS]);
 
   const renderContent = () => {
     if (stage === 'turnSelect') {
@@ -60,7 +35,7 @@ const SelectRenderContent = ({ navigation }) => {
   return (
     <GameTemplate
       subtitleText={subtitleText}
-      loading={loading}
+      loading={loading || ttsLoading}
       renderContent={renderContent}
       navigation={navigation}
     />
